qa: move more complex tools into dedicated namespaces
