+++ src/ConfigProvider.php
@@ -9,12 +9,12 @@ use Laminas\ServiceManager\Command\AheadOfTimeFactoryCreatorCommandFactory;
 use Laminas\ServiceManager\Command\ConfigDumperCommand;
 use Laminas\ServiceManager\Command\FactoryCreatorCommand;
 use Laminas\ServiceManager\Factory\InvokableFactory;
-use Laminas\ServiceManager\Tool\AheadOfTimeFactoryCompilerFactory;
-use Laminas\ServiceManager\Tool\AheadOfTimeFactoryCompilerInterface;
+use Laminas\ServiceManager\Tool\AheadOfTimeFactoryCompiler\AheadOfTimeFactoryCompilerFactory;
+use Laminas\ServiceManager\Tool\AheadOfTimeFactoryCompiler\AheadOfTimeFactoryCompilerInterface;
 use Laminas\ServiceManager\Tool\ConfigDumperFactory;
 use Laminas\ServiceManager\Tool\ConfigDumperInterface;
-use Laminas\ServiceManager\Tool\ConstructorParameterResolver;
-use Laminas\ServiceManager\Tool\ConstructorParameterResolverInterface;
+use Laminas\ServiceManager\Tool\ConstructorParameterResolver\ConstructorParameterResolver;
+use Laminas\ServiceManager\Tool\ConstructorParameterResolver\ConstructorParameterResolverInterface;
 use Laminas\ServiceManager\Tool\FactoryCreatorFactory;
 use Laminas\ServiceManager\Tool\FactoryCreatorInterface;
 use Symfony\Component\Console\Command\Command;
