+++ src/Tool/FactoryCreatorFactory.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Laminas\ServiceManager\Tool;
 
+use Laminas\ServiceManager\Tool\ConstructorParameterResolver\ConstructorParameterResolverInterface;
 use Psr\Container\ContainerInterface;
 
 /**
